refactor: id 타입 수정, 가이드 상품 생성 수정
- id 타입 int -> Long
- DB에 테이블 생성 시 SRID 명시
- ZonedDateTime 타입 JsonFormat 사용
    - @NoArgsConstructor, @AllArgsConstructor 없으면 JsonFormat 작동 X
- CreateGuideProductRequest -> GuideProductRequest로 이름 변경
- 위치 검증 로직에서 위도 경도 범위 수정

+++ src/main/java/com/swygbro/trip/backend/domain/guideProduct/api/GuideProductController.java
@@ -74,7 +74,7 @@ public class GuideProductController {
                     )
             )
     )
-    public GuideProductDto getGuideProduct(@PathVariable int productId) {
+    public GuideProductDto getGuideProduct(@PathVariable Long productId) {
         return guideProductService.getProduct(productId);
     }
 }
